Testability fix; don't share Options between instances of a client
When more than one instance of the client was created in the program,
they would share values of some options (the last instantiated one would
reconfigure the previously created instances).

This was not a problem for common usage, because there was only one
instance. Automated tests may create multiple instances without exiting
the JVM in between.