Change formatters to produce a data structure, then print that later
Previously, the client was directly building an output string right away
and then printed it. Now, it builds a `Map<String, Object>`` to represent
the message, then converts that.

The advantage is that the conversion method may later be made pluggable,
to output either Python literals (repr), or JavaScript literals (JSON).