Catch SocketException due to AMQ-6956, broker closes ssl socket 'too fast'
Java calls checkWrite() on a socket once after all data is written to it.
Upon receiving the shutdown command, the broker closes its end of connection.
If the socket is already closed at the point of the final check, we get
JMSException from Connection#close(). This is harmless and must be ignored.