NEW(artemis-jms-client) specialized AccSenderClient
In client failover scenarios in HA, blocking client (sync send) gets unblocked
by throwing an exception from broker. Client is unable to reconnect
again. Retransmission is not working at all.

As a fix, we catch such unblocked blocked call exception and resend the
message again.

This is valid only for AccSenderClient. Other clients are not affected.