Add a Mockito based test
Mockito is a mock/spy generation framework. It makes it easier to
replace some objects with test doubles (mocks, spies), then run a program
in a test, and assert about behavior of the code (what methods of the
mocks/spies were called, with what arguments, ...)

This is similar to Byteman, except that the code being tested needs
to be written in a way compatible with mocking. (It must be possible to
pass the mocked objects as constructor/method parameters. Instance
created using `new` cannot be mocked by Mockito.)

Thinking of qe clis as a program that transforms a set of cli options
into JMS API calls, we can use this to check that a given option line
results in the correct api calls. These tests should be faster to run
than end-to-end send/receive using real broker. Embedded or otherwise.