fix for Compile tests always, run only !external tests unless -Ptests is specified
First time, this commit did not work, reported and applied a workaround
https://issues.apache.org/jira/browse/SUREFIRE-1577.

Also, tags on methods maybe are not inherited? So sprinkled the tag
more widely for good measure.

Original commit message:

Tests were only compiled if -Ptests was specified. This caused weird
errors when switching from `mvn` command to intellij ide and back.

Also, it decreased utility of having the tests.

With this change, tests that are cheap to run (do not require external
broker running on the same machine) are being run.