Change pom structure in cli-qpid-jms (add parent and subprojects)
There are three variants we can choose to build cli-qpid-jms
- without tcnative openssl bindings
- with boringssl static library
- with ability to use openssl in the system (Fedora, RHEL, CentOS)

This change adds the two subprojects for the two tcnative builds,
and introduces a parent project (because parent must have packaging=pom,
so it is not possible to reuse existing project, with packaging jar or
bundle).

Also, dependencies on junit-pioneer and on the tcnative libs are added.

Also, cli-qpid-jms is modified to build test-jar.