Update protonj2 version and add staging repository
This commit updates protonj2 version from 1.0.0-M15 to 1.0.0-M16. Additionally, a new repository URL is added to the repositories list. The protonj2 version is updated to avail the newer features and to fix potential bugs. Adding the repository URL to project POM offers a location to download dependency during the build.